backend/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.1.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
keyring==25.2.1
keyrings.alt==5.0.1
more-itertools==10.3.0
protobuf==5.27.1
psycopg2-binary==2.9.9
# my fork of pyicloud
# TODO: change url to specific tag or branch rather than raw commit hash
pyicloud @ git+https://github.com/G-Town/pyicloud@2c3db04acc2e241d5c19029eab638fd792174b99
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.32.3
sqlparse==0.5.0
style==1.1.0
tzdata==2024.1
tzlocal==5.2
update==0.0.1
urllib3==2.2.1
